requirements.txt
Mastodon.py
tweepy
psycopg2